+++ app/Http/Controllers/Admin/Career/JobBoardController.php
@@ -8,8 +8,14 @@ use App\Http\Requests\Career\JobBoardUpdateRequest;
 use App\Models\Career\JobBoard;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\Rule;
 use Illuminate\View\View;
+use Illuminate\Support\Str;
 
+/**
+ *
+ */
 class JobBoardController extends BaseController
 {
     /**
@@ -30,25 +36,49 @@ class JobBoardController extends BaseController
 
     /**
      * Show the form for creating a new job board.
+     *
+     * @param Request $request
+     * @return View
      */
-    public function create(): View
+    public function create(Request $request): View
     {
-        return view('admin.career.job-board.create');
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can add job board entries.');
+        }
+
+        $referer = $request->headers->get('referer');
+
+        return view('admin.career.job-board.create', compact('referer'));
     }
 
     /**
      * Store a newly created job board in storage.
+     *
+     * @param JobBoardStoreRequest $request
+     * @return RedirectResponse
      */
     public function store(JobBoardStoreRequest $request): RedirectResponse
     {
-        JobBoard::create($request->validated());
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can add job board entries.');
+        }
 
-        return redirect()->route('admin.career.job-board.index')
-            ->with('success', 'Cover letter created successfully.');
+        $jobBoard =JobBoard::create($request->validated());
+
+        if (!empty($referer)) {
+            return redirect(str_replace(config('app.url'), '', $referer))
+                ->with('success', $jobBoard->name . ' created successfully.');
+        } else {
+            return redirect()->route('admin.career.job-board.index')
+                ->with('success', $jobBoard->name . '  created successfully.');
+        }
     }
 
     /**
      * Display the specified job board.
+     *
+     * @param JobBoard $jobBoard
+     * @return View
      */
     public function show(JobBoard $jobBoard): View
     {
@@ -57,31 +87,75 @@ class JobBoardController extends BaseController
 
     /**
      * Show the form for editing the specified job board.
+     *
+     * @param JobBoard $jobBoard
+     * @param Request $request
+     * @return View
      */
-    public function edit(JobBoard $jobBoard): View
+    public function edit(JobBoard $jobBoard, Request $request): View
     {
-        return view('admin.career.job-board.edit', compact('jobBoard'));
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can edit job board entries.');
+        }
+
+        $referer = $request->headers->get('referer');
+
+        return view('admin.career.job-board.edit', compact('jobBoard', 'referer'));
     }
 
     /**
      * Update the specified job board in storage.
+     *
+     * @param JobBoardUpdateRequest $request
+     * @param JobBoard $jobBoard
+     * @return RedirectResponse
      */
     public function update(JobBoardUpdateRequest $request, JobBoard $jobBoard): RedirectResponse
     {
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can update job board entries.');
+        }
+
+        // Validate the posted data and generated slug.
+        $validatedData = $request->validated();
+        $request->merge([ 'slug' => Str::slug($validatedData['name']) ]);
+        $request->validate(['slug' => [ Rule::unique('posts', 'slug') ] ]);
         $jobBoard->update($request->validated());
 
+        $referer = $request->input('referer');
+
+        if (!empty($referer)) {
+            return redirect(str_replace(config('app.url'), '', $referer))
+                ->with('success', $jobBoard->name . ' updated successfully.');
+        } else {
         return redirect()->route('admin.career.job-board.index')
-            ->with('success', 'Cover letter updated successfully');
+            ->with('success', $jobBoard->name . ' updated successfully');
+        }
     }
 
     /**
      * Remove the specified job board from storage.
+     *
+     * @param JobBoard $jobBoard
+     * @param Request $request
+     * @return RedirectResponse
      */
-    public function destroy(JobBoard $jobBoard): RedirectResponse
+    public function destroy(JobBoard $jobBoard, Request $request): RedirectResponse
     {
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can delete job board entries.');
+        }
+
         $jobBoard->delete();
 
-        return redirect()->route('admin.job-board.link.index')
-            ->with('success', 'Cover letter deleted successfully');
+        $referer = $request->input('referer');
+
+        if (!empty($referer)) {
+            return redirect(str_replace(config('app.url'), '', $referer))
+                ->with('success', $jobBoard->name . ' deleted successfully.');
+        } else {
+            return redirect()->route('admin.job-board.link.index')
+                ->with('success', $jobBoard->name . ' deleted successfully');
+        }
     }
 }
